Add import/export mode getters
