+++ kolibri/plugins/device_management/assets/src/constants.js
@@ -38,3 +38,14 @@ export const TransferTypes = {
   REMOTEIMPORT: 'remoteimport',
   LOCALEXPORT: 'localexport',
 };
+
+export const ContentWizardErrors = {
+  INVALID_PARAMETERS: 'INVALID_PARAMETERS',
+  CHANNEL_NOT_FOUND_ON_SERVER: 'CHANNEL_NOT_FOUND_ON_SERVER',
+  CHANNEL_NOT_FOUND_ON_DRIVE: 'CHANNEL_NOT_FOUND_ON_DRIVE',
+  CHANNEL_NOT_FOUND_ON_STUDIO: 'CHANNEL_NOT_FOUND_ON_STUDIO',
+  KOLIBRI_STUDIO_UNAVAILABLE: 'KOLIBRI_STUDIO_UNAVAILABLE',
+  DRIVE_IS_NOT_WRITEABLE: 'DRIVE_IS_NOT_WRITEABLE',
+  DRIVE_NOT_FOUND: 'DRIVE_NOT_FOUND',
+  TRANSFER_IN_PROGRESS: 'TRANSFER_IN_PROGRESS',
+};
