+++ kolibri/plugins/device_management/assets/src/state/actions/contentTransferActions.js
@@ -1,6 +1,6 @@
 import { ChannelResource, TaskResource } from 'kolibri.resources';
-import { taskList, wizardState } from '../getters';
-import { TaskStatuses, TransferTypes } from '../../constants';
+import { taskList, wizardState, inLocalImportMode, inRemoteImportMode } from '../getters';
+import { TaskStatuses } from '../../constants';
 
 export const ErrorTypes = {
   CONTENT_DB_LOADING_ERROR: 'CONTENT_DB_LOADING_ERROR',
@@ -13,17 +13,19 @@ export const ErrorTypes = {
  *
  */
 export function downloadChannelMetadata(store) {
-  const { transferType, transferredChannel, selectedDrive } = wizardState(store.state);
+  const { transferredChannel, selectedDrive } = wizardState(store.state);
   let promise;
-  if (transferType === TransferTypes.LOCALIMPORT) {
+  if (inLocalImportMode(store.state)) {
     promise = TaskResource.startDiskChannelImport({
       channel_id: transferredChannel.id,
       drive_id: selectedDrive.id,
     });
-  } else if (transferType === TransferTypes.REMOTEIMPORT) {
+  } else if (inRemoteImportMode(store.state)) {
     promise = TaskResource.startRemoteChannelImport({
       channel_id: transferredChannel.id,
     });
+  } else {
+    return Error('Channel Metadata is only downloaded when importing');
   }
   promise = promise.catch(() => Promise.reject({ errorType: ErrorTypes.CONTENT_DB_LOADING_ERROR }));
 
@@ -43,37 +45,32 @@ export function downloadChannelMetadata(store) {
     });
 }
 
-const combinePks = nodes => nodes.map(({ pk }) => pk);
-
 /**
  * Starts a Task that transfers Channel ContentNodes to/from a drive
  *
  */
 export function transferChannelContent(store) {
   let promise;
-  const { transferType, transferredChannel, selectedDrive, nodesForTransfer } = wizardState(
-    store.state
-  );
+  const combinePks = nodes => nodes.map(({ pk }) => pk);
+  const { transferredChannel, selectedDrive, nodesForTransfer } = wizardState(store.state);
   const params = {
     channel_id: transferredChannel.id,
     node_ids: combinePks(nodesForTransfer.included),
     exclude_node_ids: combinePks(nodesForTransfer.omitted),
   };
-  switch (transferType) {
-    case TransferTypes.REMOTEIMPORT:
-      promise = TaskResource.startRemoteContentImport(params);
-      break;
-    case TransferTypes.LOCALIMPORT:
-      promise = TaskResource.startDiskContentImport({
-        ...params,
-        drive_id: selectedDrive.id,
-      });
-      break;
-    case TransferTypes.LOCALEXPORT:
-      promise = TaskResource.startDiskContentExport({
-        ...params,
-        drive_id: selectedDrive.id,
-      });
+
+  if (inRemoteImportMode(store.state)) {
+    promise = TaskResource.startRemoteContentImport(params);
+  } else if (inLocalImportMode(store.state)) {
+    promise = TaskResource.startDiskContentImport({
+      ...params,
+      drive_id: selectedDrive.id,
+    });
+  } else {
+    promise = TaskResource.startDiskContentExport({
+      ...params,
+      drive_id: selectedDrive.id,
+    });
   }
   return promise;
 }
